refactor(listener): split unix socket helpers
