+++ CMakeLists.txt
@@ -56,6 +56,7 @@ set(COMMON_SOURCES
   src/status.c
   src/connection.c
   src/worker.c
+  src/unix_socket.c
   src/buffer_pool.c
   src/zerocopy.c
   src/m3u.c
